# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/endorsements:
    get:
      operationId: listAdminEndorsements
      tags: [editorial]
      summary: List all endorsements for administrators
      responses:
        "200":
          description: All endorsements ordered for administration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEndorsement"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createAdminEndorsement
      tags: [editorial]
      summary: Create an endorsement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndorsementInput"
      responses:
        "201":
          description: Endorsement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEndorsement"
        "400":
          description: Invalid endorsement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/endorsements/order:
    patch:
      operationId: reorderAdminEndorsements
      tags: [editorial]
      summary: Set endorsement display order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndorsementOrderInput"
      responses:
        "200":
          description: Reordered endorsements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEndorsement"
        "400":
          description: Invalid endorsement order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/endorsements/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getAdminEndorsement
      tags: [editorial]
      summary: Get an endorsement by ID
      responses:
        "200":
          description: Endorsement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEndorsement"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Endorsement not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAdminEndorsement
      tags: [editorial]
      summary: Update, approve, or archive an endorsement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndorsementUpdate"
      responses:
        "200":
          description: Updated endorsement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEndorsement"
        "400":
          description: Invalid endorsement update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Endorsement not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAdminEndorsement
      tags: [editorial]
      summary: Permanently delete an endorsement by ID
      responses:
        "204":
          description: Endorsement deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Endorsement not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /enquiries:
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a website enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "200":
          description: Generic success response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquirySubmission"
        "400": { description: Invalid enquiry, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "429": { description: Too many submissions, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /admin/enquiries:
    get:
      operationId: listAdminEnquiries
      tags: [enquiries]
      summary: List enquiries for administrators
      responses:
        "200":
          description: Enquiries and summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEnquiryList"
        "401": { description: Authentication required }
        "403": { description: Administrator role required }
  /admin/enquiries/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getAdminEnquiry
      tags: [enquiries]
      summary: Get an enquiry
      responses:
        "200":
          description: Enquiry detail
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminEnquiry" } } }
        "404": { description: Enquiry not found }
    patch:
      operationId: updateAdminEnquiry
      tags: [enquiries]
      summary: Update enquiry review status
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnquiryUpdate" }
      responses:
        "200":
          description: Updated enquiry
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminEnquiry" } } }
        "400": { description: Invalid update }
        "404": { description: Enquiry not found }
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    EndorsementStatus:
      type: string
      enum: [draft, approved, archived]
    Endorsement:
      type: object
      required: [id, name, quote, displayOrder]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        title:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        quote:
          type: string
          minLength: 1
        photoUrl:
          type: ["string", "null"]
          format: uri
        photoMediaId:
          type: ["string", "null"]
          format: uuid
        language:
          type: string
          enum: [english, french, bilingual]
        sourceUrl:
          type: ["string", "null"]
          format: uri
        displayOrder:
          type: integer
    AdminEndorsement:
      allOf:
        - $ref: "#/components/schemas/Endorsement"
        - type: object
          required: [status, createdAt, updatedAt]
          properties:
            status:
              $ref: "#/components/schemas/EndorsementStatus"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            approvedBy:
              type: ["string", "null"]
            approvedAt:
              type: ["string", "null"]
              format: date-time
            language:
              type: string
              enum: [english, french, bilingual]
            verificationNote:
              type: ["string", "null"]
            verified:
              type: boolean
            photoMediaId:
              type: ["string", "null"]
              format: uuid
    EndorsementInput:
      type: object
      required: [name, quote, status]
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        quote:
          type: string
          minLength: 1
        photoUrl:
          type: ["string", "null"]
          format: uri
        sourceUrl:
          type: ["string", "null"]
          format: uri
        language:
          type: string
          enum: [english, french, bilingual]
        verificationNote:
          type: string
          nullable: true
        verified:
          type: boolean
        photoMediaId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/EndorsementStatus"
        displayOrder:
          type: integer
          minimum: 0
    EndorsementUpdate:
      minProperties: 1
      type: object
      properties:
        approvalConfirmation:
          type: string
        name:
          type: string
          minLength: 1
        title:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        quote:
          type: string
          minLength: 1
        photoUrl:
          type: ["string", "null"]
          format: uri
        sourceUrl:
          type: ["string", "null"]
          format: uri
        language:
          type: string
          enum: [english, french, bilingual]
        verificationNote:
          type: string
          nullable: true
        verified:
          type: boolean
        photoMediaId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/EndorsementStatus"
        displayOrder:
          type: integer
          minimum: 0
...
    EndorsementOrderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [id, displayOrder]
            properties:
              id:
                type: string
                format: uuid
              displayOrder:
                type: integer
                minimum: 0
    EnquiryInput:
      type: object
      additionalProperties: false
      required: [name, email, enquiryType, subject, message, language, honeypot]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email, maxLength: 320 }
        organization: { type: ["string", "null"], maxLength: 160 }
        enquiryType: { type: string, enum: [technical_assistance, strategic_advisory, academic_collaboration, policy_dialogue, other] }
        subject: { type: string, minLength: 1, maxLength: 200 }
        message: { type: string, minLength: 1, maxLength: 10000 }
        language: { type: string, enum: [english, french] }
        honeypot: { type: string, maxLength: 0 }
    EnquirySubmission:
      type: object
      additionalProperties: false
      required: [accepted]
      properties: { accepted: { type: boolean, enum: [true] } }
    EnquiryReviewStatus:
      type: string
      enum: [new, in_progress, resolved]
    EnquiryDeliveryStatus:
      type: string
      enum: [pending, sent, failed]
    AdminEnquiry:
      type: object
      required: [id, name, email, enquiryType, subject, message, language, reviewStatus, deliveryStatus, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        organization: { type: ["string", "null"] }
        enquiryType: { type: string }
        subject: { type: string }
        message: { type: string }
        language: { type: string }
        reviewStatus: { $ref: "#/components/schemas/EnquiryReviewStatus" }
        deliveryStatus: { $ref: "#/components/schemas/EnquiryDeliveryStatus" }
        providerId: { type: ["string", "null"] }
        providerError: { type: ["string", "null"] }
        deliveryAttemptedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EnquiryUpdate:
      type: object
      additionalProperties: false
      required: [reviewStatus]
      properties:
        reviewStatus: { $ref: "#/components/schemas/EnquiryReviewStatus" }
    AdminEnquiryList:
      type: object
      required: [items, counts]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AdminEnquiry" }
        counts:
          type: object
          required: [total, new, inProgress, resolved, deliveryFailed]
          properties:
            total: { type: integer }
            new: { type: integer }
            inProgress: { type: integer }
            resolved: { type: integer }
            deliveryFailed: { type: integer }